requirements.txt
matplotlib==3.6.3
pandas==1.4.2
Pillow==9.4.0
seaborn==0.11.2
streamlit==1.16.0
